requirements.txt
asgiref==3.4.1
Django==2.2.17
djangorestframework==3.12.4
pytz==2021.3
sqlparse==0.4.2
Pillow==8.4.0